All the files are refactored to be in the format of Pascals Case
