+++ client/components/ImageDisplay.jsx
